working renderer

+++ src/gamedata/color.rs
@@ -1,9 +1,9 @@
 #[derive(Debug, Clone, PartialEq)]
 pub struct Color {
-    pub r: f64,
-    pub g: f64,
-    pub b: f64,
-    pub a: f64,
+    pub r: f32,
+    pub g: f32,
+    pub b: f32,
+    pub a: f32,
 }
 
 impl Default for Color {
